pset3/requirements.txt
pygame==2.6.1
PyOpenGL==3.1.0
opencv-python==4.9.0.80
matplotlib
numpy==1.24.1
scikit-image
Pillow
tqdm
-e .